video_audio_align/requirements.txt
openai-whisper
ffmpeg-python
Pillow
numpy
moviepy
